+++ types/ramda/test/product-tests.ts
@@ -1,0 +1,5 @@
+import * as R from 'ramda';
+
+() => {
+  R.product([2, 4, 6, 8, 100, 1]); // => 38400
+};
