+++ types/ramda/test/props-tests.ts
@@ -1,0 +1,5 @@
+import * as R from 'ramda';
+
+() => {
+  R.props(['x', 'y'], { x: 1, y: 2 }); // => [1, 2]
+};
